fix(UI): Move the view planes button into the volume UI
This button is relevant only with the volume rendering view mode. Only display it with the other
relevant buttons.

Suggested-by: Alexis Girault <alexis.girault@kitware.com>
